Create README.md
THE ROVERS CLUB WEBSITE PROJECT
THE ROVERS is a new hiking group that has just started with around 50 participants.
It Has been operating in the Kisii region for around four months.
The group intends to expand its activities from hiking to other various activities i.e. Camping, partying, Game-Drives, Volunteer works, sports, road trips, etc.
All to expand its members and by doing so, to reach countrywide.
To make this a success, a web application plays an essential role to achieve this, mainly for advertising the ROVERS, and to reach many people as possible
Anyone who meets the minimum requirements of membership can register with the club through the web application, full details of the member are required.
The user can navigate through various sections
1.	Events
2.	Gallery
3.	History
4.	Account
5.	Suggestions.
6.	HOME
In the Events menu:
Detailed information of the event is given: Venue of the event, Date, and Time In addition to the venue (a navigation map is available to make it easy to reach to the event and gives the route for hiking/touring). Images of the scenery are displayed so the user can have an idea of the place they’re going to.
The navigation offers the route and distance of the hiking event
And also, the fun activities to be done in the event.
It also shows the previous events that had taken place and reactions from the participants.
In the Gallery menu:
It stores memorable group pictures of the club in various tours and hikes.
About:
The description of the group, its activities, location, usernames of the officials, and usernames of the members
Suggestions:
Every member has the right to suggest an event and this is done through voting for the suggested event's venue.
Account:
signing in and out.
This includes the user details
The history of your participation in the ROVERS (number of events you have participated in).
Credits are also given to members considering their efforts in participating in the hiking events, Bringing members to the club. This may be used as vouchers for any upcoming events that require fee payment.
The participant has a right to react to the event participated.